Malnick/generate checks

+++ action/generate.go
@@ -5,14 +5,23 @@ import (
 	"crypto/rsa"
 	"crypto/x509"
 	"encoding/pem"
+	"errors"
 	"fmt"
 	"io/ioutil"
+	"os"
 
 	log "github.com/Sirupsen/logrus"
 	"github.com/malnick/cryptorious/config"
 )
 
-// GenerateKeys() creates public private keys for a $USER
+func checkFileExists(path string) bool {
+	if _, err := os.Stat(path); err == nil {
+		return true
+	}
+	return false
+}
+
+// GenerateKeys creates public private keys for a $USER
 func GenerateKeys(c config.Config) error {
 	privPath := c.PrivateKeyPath
 	pubPath := c.PublicKeyPath
@@ -27,6 +36,10 @@ func GenerateKeys(c config.Config) error {
 		Bytes: x509.MarshalPKCS1PrivateKey(privatekey),
 	})
 
+	if checkFileExists(privPath) {
+		log.Warnf("%s exists, please manually remove to proceed.", privPath)
+		return errors.New("Will not overwrite existing private key path.")
+	}
 	if err := ioutil.WriteFile(privPath, privBytes, 0600); err != nil {
 		return err
 	}
@@ -43,6 +56,10 @@ func GenerateKeys(c config.Config) error {
 		Bytes: ansipub,
 	})
 
+	if checkFileExists(pubPath) {
+		log.Warnf("%s exists, please manually remove to proceed.", pubPath)
+		return errors.New("Will not overwrite existing public key path.")
+	}
 	if err := ioutil.WriteFile(pubPath, pubBytes, 0644); err != nil {
 		return err
 	}
